requirements.txt
google-cloud-vision==1.0

python-telegram-bot==13.1

